+++ src/main/resources/db/migration/V2__Create_question_table.sql
@@ -1,11 +1,11 @@
 create table question
 (
-	id INT auto_increment,
+	id BIGINT auto_increment,
 	title VARCHAR(50),
 	description TEXT,
 	gmt_create BIGINT,
 	gmt_modified BIGINT,
-	creator INT,
+	creator BIGINT,
 	comment_count INT default 0,
 	view_count INT default 0,
 	thumb_count INT default 0,
